move train tools to top-level module
